Add context to App
